+++ src/app/components/historial/historial.component.ts
