+++ src/app/components/inicio/inicio.component.html
@@ -1,0 +1,5 @@
+<app-navbar></app-navbar>
+<h1  style="text-align:center;">
+  Proyectos Activos
+</h1>
+<br/>
