--- src/app/inicio/inicio.component.html
@@ -1,54 +1,0 @@
-<app-navbar></app-navbar>
-<h1  style="text-align:center;">
-  Proyectos Activos
-</h1>
-<br>
-<!--
-<div class="container" id="grid" style="color:white"///////////////////////////////////////////>
-  <div class="row">
-    <div class="col-sm">
-      Proyecto #1
-    </div>
-    <div class="col-sm">
-      Proyecto #2
-    </div>
-    <div class="col-sm">
-      Proyecto #3
-    </div>
-  </div>
-</div>
-<br>
-<div class="container" id="grid" style="color:white">
-  <div class="row">
-    <div class="col-sm">
-      Proyecto #4
-    </div>
-    <div class="col-sm">
-      Proyecto #5
-    </div>
-    <div class="col-sm">
-      Proyecto #6
-    </div>
-  </div>
-</div>
--->
-<br/>
-<div *ngFor="let proyecto of proyectos">
-  <div id="card" class="card" style="width: 18rem;">
-      <div class="card-body">
-        <h5 class="card-title">{{proyecto.titulo}}</h5>
-        <h6 class="card-subtitle mb-2 text-muted">{{proyecto.idCliente}}</h6>
-        <p class="card-text">{{proyecto.key}}</p>
-        <p class="card-text">{{proyecto.estado}}</p>
-        <a href="#" class="card-link">Ver Detalles</a>
-      </div>
-    </div>    
-  <br/>
-</div>
-
-<div
-  infiniteScroll
-  [infiniteScrollDistance]="2"
-  [infiniteScrollThrottle]="1500"
-  (scrolled)="onScroll()">
-</div>
