+++ src/app/model/Equipo.ts
